added fps counter
a large random world is displayed now

+++ include/chunk.h
@@ -7,7 +7,7 @@
 #include "mesh.h"
 
 typedef struct {
-	int points;
+	long points;
 	block_t *data;
 } chunk_t;
